+++ src/util.js
@@ -9,9 +9,6 @@ const getRandomArrayElement = (items) => (
   items[Math.floor(Math.random() * items.length)]
 );
 
-const DATE_FORMAT_DATE = 'DD MMM';
-const DATE_FORMAT_TIME = 'HH:mm';
-
 const isEscapeKey = (evt) => (evt.key === 'Escape' || evt.key === 'Esc');
 
-export {getRandomNumber, getRandomArrayElement, DATE_FORMAT_DATE, DATE_FORMAT_TIME, isEscapeKey};
+export {getRandomNumber, getRandomArrayElement, isEscapeKey};
